task4_search_images_with_text/requirements.txt
torch==2.3.0
torchvision==0.18.0
transformers==4.41.2
pillow==10.3.0
pandas==2.2.2
matplotlib==3.9.0
scikit-learn==1.5.0